requirements.txt
decorator
flask_limiter
flask_restful
python-vlc
pyyaml
requests
